Build scripts grouper kernel:
Clean the kernel first.  Since updating to SaberMod's daily toolchains,
will likely change on a daily basis.  The kernel source
does not detect changes in the toolchain binaries, and therefore does
not rebuild any parts of the kernel source.  Because of this, should
be building a fresh kernel each time to ensure new code is generated
using the new toolchain binaries.

Signed-off-by:  Paul Beeler <sparksco@gmail.com>